requirements.txt
requests
mutagen
click
tabulate
wcwidth
pdfkit
m3u8
ffmpeg-python
